fix: Moving RatingScore and StarIcons to a new component file
