+++ src/components/index.jsx
@@ -1,5 +1,7 @@
 import Breadcrumbs from './Breadcrumbs'
+import StarIcon from './icons/StarIcon'
+import RatingScore from './RatingScore'
 import Searchbar from './Searchbar'
 import TableData from './TableData'
 
-export { Breadcrumbs, Searchbar, TableData }
+export { Breadcrumbs, Searchbar, TableData, RatingScore, StarIcon }
